+++ src/main/java/com/ing/data/cassandra/jdbc/SpecialCommands.java
@@ -15,7 +15,9 @@
 
 package com.ing.data.cassandra.jdbc;
 
+import com.datastax.oss.driver.api.core.ConsistencyLevel;
 import com.datastax.oss.driver.api.core.DefaultConsistencyLevel;
+import com.datastax.oss.driver.api.core.cql.ResultSet;
 import com.ing.data.cassandra.jdbc.utils.SpecialCommandsUtil;
 
 import javax.annotation.Nullable;
@@ -36,6 +38,7 @@ import static com.ing.data.cassandra.jdbc.utils.SpecialCommandsUtil.buildSpecial
  *     Supported commands:
  *     <ul>
  *         <li>{@code CONSISTENCY [level]}</li>
+ *         <li>{@code SERIAL CONSISTENCY [level]}</li>
  *         </ul>
  * </p>
  */
@@ -58,6 +61,16 @@ public final class SpecialCommands {
          */
         com.datastax.oss.driver.api.core.cql.ResultSet execute(CassandraStatement statement,
                                                                String cql) throws SQLException;
+    }
+
+    /**
+     * Executor returning an empty result set.
+     */
+    public static class NoOpExecutor implements SpecialCommandExecutor {
+        @Override
+        public ResultSet execute(final CassandraStatement statement, final String cql) throws SQLException {
+            return buildEmptyResultSet();
+        }
     }
 
     /**
@@ -92,7 +105,11 @@ public final class SpecialCommands {
             final CassandraConnection connection = (CassandraConnection) statement.getConnection();
 
             if (this.levelParameter != null) {
-                connection.setConsistencyLevel(DefaultConsistencyLevel.valueOf(this.levelParameter));
+                final ConsistencyLevel newConsistencyLevel = DefaultConsistencyLevel.valueOf(this.levelParameter);
+                connection.setConsistencyLevel(newConsistencyLevel);
+                // Also update the consistency level of the current statement (in case this one contains several
+                // statements to execute, especially after the consistency level update).
+                statement.setConsistencyLevel(newConsistencyLevel);
                 return buildEmptyResultSet();
             } else {
                 final String currentLevel = connection.getConsistencyLevel().name();
@@ -103,6 +120,62 @@ public final class SpecialCommands {
                 return buildSpecialCommandResultSet(
                     new ColumnDefinitions.Definition[]{
                         ColumnDefinitions.Definition.buildDefinitionInAnonymousTable("consistency_level", TEXT)
+                    },
+                    Collections.singletonList(
+                        Collections.singletonList(currentLevelAsBytes)
+                    )
+                );
+            }
+        }
+    }
+
+    /**
+     * Executor for serial consistency level special commands.
+     * <p>
+     *     {@code SERIAL CONSISTENCY [level]}: if {@code level} is specified, set the current serial consistency level
+     *     of the connection to the given value, otherwise return a return a result set with a single row containing the
+     *     current serial consistency level in a column {@code serial_consistency_level}.
+     * </p>
+     */
+    public static class SerialConsistencyLevelExecutor implements SpecialCommandExecutor {
+
+        private String levelParameter = null;
+
+        /**
+         * Constructor.
+         *
+         * @param levelParameter The parameter {@code level} of the command, if defined.
+         */
+        public SerialConsistencyLevelExecutor(@Nullable final String levelParameter) {
+            if (levelParameter != null) {
+                this.levelParameter = levelParameter.toUpperCase(Locale.ENGLISH);
+            }
+        }
+
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        public com.datastax.oss.driver.api.core.cql.ResultSet execute(final CassandraStatement statement,
+                                                                      final String cql) throws SQLException {
+            final CassandraConnection connection = (CassandraConnection) statement.getConnection();
+
+            if (this.levelParameter != null) {
+                final ConsistencyLevel newConsistencyLevel = DefaultConsistencyLevel.valueOf(this.levelParameter);
+                connection.setSerialConsistencyLevel(newConsistencyLevel);
+                // Also update the serial consistency level of the current statement (in case this one contains several
+                // statements to execute, especially after the serial consistency level update).
+                statement.setSerialConsistencyLevel(newConsistencyLevel);
+                return buildEmptyResultSet();
+            } else {
+                final String currentLevel = connection.getSerialConsistencyLevel().name();
+
+                // Create a result set with a single row containing the current consistency level value in a column
+                // 'serial_consistency_level'.
+                final ByteBuffer currentLevelAsBytes = bytes(currentLevel);
+                return buildSpecialCommandResultSet(
+                    new ColumnDefinitions.Definition[]{
+                        ColumnDefinitions.Definition.buildDefinitionInAnonymousTable("serial_consistency_level", TEXT)
                     },
                     Collections.singletonList(
                         Collections.singletonList(currentLevelAsBytes)