+++ CHANGELOG.md
@@ -4,11 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [4.14.0] - Unreleased
 ### Added
 - Add support for IPv6 addresses in JDBC URL (see PR [#62](https://github.com/ing-bank/cassandra-jdbc-wrapper/pull/62)).
 - Add implementation for the method `CassandraPreparedStatement.setArray(int, Array)`.
-### Changed
+- Add a parameter `serialconsistency` to specify the default serial consistency level used by the connection.
+- Add support for the special CQL command `SERIAL CONSISTENCY [level]` in `CassandraStatement`.
+### Fixed
 - Do not try to register codecs again (if already done previously) on a pre-existing session to avoid warnings in logs.
 
 ## [4.13.1] - 2024-09-04
@@ -306,6 +308,7 @@ For this version, the changelog lists the main changes comparatively to the late
 - Fix logs in `CassandraConnection` constructor.
 
 [original project]: https://github.com/adejanovski/cassandra-jdbc-wrapper/
+[4.14.0]: https://github.com/ing-bank/cassandra-jdbc-wrapper/compare/v4.13.1...v4.14.0
 [4.13.1]: https://github.com/ing-bank/cassandra-jdbc-wrapper/compare/v4.13.0...v4.13.1
 [4.13.0]: https://github.com/ing-bank/cassandra-jdbc-wrapper/compare/v4.12.0...v4.13.0
 [4.12.0]: https://github.com/ing-bank/cassandra-jdbc-wrapper/compare/v4.11.1...v4.12.0